extract kinesis
